requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
djangorestframework==3.10.2
idna==2.8
pkg-resources==0.0.0
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
